Use container /tmp as HOME.
Currently the users real home is passed into the container and used. This means that the git command to mark everything as safe is being set in the real user gitconfig which is a security issue.

Using the temp folder in the container ensures changes like this are ephemeral and limited to the build environment.